pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tiny-rag-warmup"
version = "0.1.0"
description = "RAG pipeline for financial 10-K documents"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "structlog",
    "tenacity",
    "tiktoken",
    "sentence-transformers",
    "pgvector",
    "psycopg2-binary",
    "ollama",
    "fastapi",
    "uvicorn",
    "pandas",
    "openpyxl",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
